+++ src/core/scheduler.h
@@ -15,10 +15,11 @@
 #ifndef CORE_SCHEDULER_H_
 #define CORE_SCHEDULER_H_
 
+#include <algorithm>
 #include <chrono>
 #include <functional>
-#include <set>
 #include <string>
+#include <unordered_map>
 #include <vector>
 #include "core/operation/operation.h"
 
@@ -45,18 +46,40 @@ class Scheduler {
   /// This function returns the numer of simulated steps (=iterations).
   uint64_t GetSimulatedSteps() const;
 
-  void AddOperation(Operation* operation);
+  void ScheduleOp(Operation* op);
 
-  /// Remove an operation. However, some operations are protected and cannot
-  /// be removed. \see protected_operations_
-  /// A request to remove a proteced operation is ignored.
-  void RemoveOperation(const std::string& op_name);
+  void UnscheduleOp(Operation* op);
 
-  /// Returns a reference to an operation. However, some operations are
-  /// protected and will not be returned. \see protected_operations_
-  /// If the operation does not exist or is protected, a nullptr will be
-  /// returned.
-  Operation* GetOperation(const std::string& op_name);
+  Operation* GetDefaultOp(const std::string& name);
+
+  /// Return a list of operations that are scheduled by default
+  std::vector<std::string> GetListOfDefaultOps() { return default_ops_; }
+
+  /// Return a list of SimObjectOperations that are scheduled
+  std::vector<std::string> GetListOfScheduledSimObjectOps() {
+    std::vector<std::string> list;
+    for (auto *op : scheduled_sim_object_ops_) {
+      list.push_back(op->name_);
+    }
+    return list;
+  }
+
+  /// Return a list of StandAloneOperations that are scheduled
+  std::vector<std::string> GetListOfScheduledStandaloneOps() {
+    std::vector<std::string> list;
+    for (auto *op : scheduled_standalone_ops_) {
+      list.push_back(op->name_);
+    }
+    return list;
+  }
+
+  void SetUpOps();
+
+  void TearDownOps();
+
+  void RunScheduledOps();
+
+  void ScheduleOps();
 
   RootAdaptor* GetRootVisualization() { return root_visualization_; }
 
@@ -74,12 +97,23 @@ class Scheduler {
   VisualizationAdaptor* visualization_ = nullptr;  //!
   RootAdaptor* root_visualization_ = nullptr;      //!
 
-  BoundSpace* bound_space_;
-  DisplacementOp* displacement_;
-  DiffusionOp* diffusion_;
-
-  std::vector<Operation*> operations_;  //!
-  std::set<std::string> protected_operations_;
+  /// List of operations that are to be added in the upcoming timestep
+  std::vector<Operation*> ops_to_add_;  //!
+  /// List of operations that are to be removed in the upcoming timestep
+  std::vector<Operation*> ops_to_remove_;  //!
+  /// List of operations that were removed from scheduling, but could be reused
+  /// later on in a simulation
+  std::vector<Operation*> unscheduled_ops_;  //!
+  /// List of operations will be executed as a stand-alone operation
+  std::vector<Operation*> scheduled_standalone_ops_;  //!
+  /// List of operations will be executed on all simulation objects
+  std::vector<Operation*> scheduled_sim_object_ops_;  //!
+  /// List of operations that are scheduled by default
+  std::vector<std::string> default_ops_;  //!
+  /// List of operations that cannot be affected by the user
+  std::vector<std::string> protected_ops_;  //!
+  /// Flag to check if scheduler has beeen initialized at least once
+  bool initialized_ = false;
 
   /// Backup the simulation. Backup interval based on `Param::backup_interval_`
   void Backup();
@@ -93,9 +127,6 @@ class Scheduler {
   // think about a better solution, because some operations are executed twice
   // if Simulate is called with one timestep.
   void Initialize();
-
-  // Decide which operations should be executed
-  std::vector<Operation*> GetScheduleOps();
 };
 
 }  // namespace bdm